pyproject.toml
[project]
name = "text-to-sql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.10.0",
    "grandalf>=0.8",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langgraph>=1.1.2",
    "omegaconf>=2.3.0",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]
